Update build notifications - ignore changes, link to console log
Summary:
While running this plugin at Khan Academy, we've seen that the change
list on build start is often unexpected (since an initiating user often
doesn't realize that it represents all changes since the last success,
or whatever algorithm is used). Also, that same change list on the build
page can be confusing, but the console log is often relevant.

To address these, we now simplify notify how the build was started, with
a more subtle link to the build page (the build display name is linked)
and an additional link to the full console log.

This is pretty hacked-in, since there isn't yet a framework in the
plugin for customizing or overriding notification messages.

Test Plan:
I ran "mvn hpi:run" to load Jenkins with the plugin. Then I created a
new empty job and checked all the options at the top to run HipChat
notification on any event, and then I checked the post-build action at
the bottom to enable HipChat notifications, then I started a build and
saw this in the logs:

  test - <a href='nulljob/test/1/'>#1</a> Started by user anonymous (<a href='nulljob/test/1/consoleFull'>Console log</a>)

It was in the context of a failure to write the message, since I didn't
actually configure Jenkins to write to HipChat.

Then, I enabled user self-signup, registered a user, and built again,
verifying that this time I saw the user name in the message:

  test - <a href='nulljob/test/2/'>#2</a> Started by user chris k (<a href='nulljob/test/2/consoleFull'>Console log</a>)

Reviewers: csilvers

Differential Revision: http://phabricator.khanacademy.org/D11678